% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbRast}
\alias{bbRast}
\title{Bioclimatic Balance}
\format{
A PackedSpatRaster with 12 monthly values of 4 variables:
\code{B}, \code{b}, \code{bc} and \code{bl}.
}
\usage{
bbRast
}
\description{
A SpatRaster containing the bioclimatic balance of the Alps.
}
\keyword{datasets}
