% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ith.R
\name{ith}
\alias{ith}
\title{Function to calculate Thornthwaite’s index}
\usage{
ith(bh)
}
\arguments{
\item{bh}{Water balance in data.frame format from watbal() function.}
}
\value{
Character, describing the humid characteristics of the climate.
}
\description{
This function calculates Thornthwaite’s index to refine the bioclimatic classification.
}
\examples{
wb <- watbal(t = rnorm(12, 18, 6), p = rnorm(12, 50, 30), lat = 35, CC = 400)
ith(wb)
}
