% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_biobal.R
\name{plotBiobal}
\alias{plotBiobal}
\title{Function to plot bioclimatic balance}
\usage{
plotBiobal(intens)
}
\arguments{
\item{intens}{bioclimatic intensities in data.frame format from bioint() function.}
}
\value{
Plot of bioclimatic balance
}
\description{
Function to plot bioclimatic balance.
}
\examples{
wb <- watbal(t = c(10, 11.5, 14, 16.5, 20, 24.5, 27.5, 28, 24.5, 19.5, 14.5, 11),
             p = c(55, 73, 84, 58, 33, 23, 2, 2, 28, 66, 94, 71), lat = 35, CC = 400)
bb <- biobal(wb, 400)
bi <- bioint(bb)
plotBiobal(bi)

}
