% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_watbal.R
\name{plotWatbal}
\alias{plotWatbal}
\title{Function to plot water balance}
\usage{
plotWatbal(bh)
}
\arguments{
\item{bh}{Water balance in data.frame format from watbal() function.}
}
\value{
Plot of water balance
}
\description{
Function to plot water balance.
}
\examples{
wb <- watbal(t = c(10, 11.5, 14, 16.5, 20, 24.5, 27.5, 28, 24.5, 19.5, 14.5, 11),
p = c(55, 73, 84, 58, 33, 23, 2, 2, 28, 66, 94, 71), lat = 35, CC = 400)
plotWatbal(wb)
}
