% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_indomb.R
\name{plotWL}
\alias{plotWL}
\title{Function to plot Walter and Lieth diagram}
\usage{
plotWL(t, p)
}
\arguments{
\item{t}{Monthly average temperature data (12 nueric values).}

\item{p}{Monthly average precipitation data (12 nueric values).}
}
\value{
Plot of Walter and Lieth diagram
}
\description{
Function to plot Walter and Lieth diagram.
}
\examples{
plotWL(t = rnorm(12, 18, 6), p = rnorm(12, 50, 30))
}
