% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_postemp.R
\name{postemp}
\alias{postemp}
\title{Function to Positive Temperature index}
\usage{
postemp(t)
}
\arguments{
\item{t}{Monthly average temperature data (12 nueric values).}
}
\value{
Positive Temperature index
}
\description{
Computes Positive Temperature index from monthly temperature.
}
\examples{
postemp(rnorm(12, 18, 6))
}
