% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preRast}
\alias{preRast}
\title{Monthly precipitation}
\format{
A PackedSpatRaster with 12 monthly values precipitation:
}
\usage{
preRast
}
\description{
A SpatRaster containing the monthly precipitation of the Alps.
}
\keyword{datasets}
