% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmpRast}
\alias{tmpRast}
\title{Monthly temperature}
\format{
A PackedSpatRaster with 12 monthly values temperature:
}
\usage{
tmpRast
}
\description{
A SpatRaster containing the monthly temperature of the Alps.
}
\keyword{datasets}
