% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wbRast}
\alias{wbRast}
\title{Water Balance}
\format{
A PackedSpatRaster with 12 monthly values of 12 variables:
  \code{temp}, \code{prec}, \code{PET}, \code{P_PET}, \code{PPA},
  \code{ST}, \code{i_ST}, \code{RET}, \code{HD}, \code{HEX}, \code{r},
  \code{rP}.
}
\usage{
wbRast
}
\description{
A SpatRaster containing the water balance of the Alps.
}
\keyword{datasets}
